demo/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "continuum-demo-api"
version = "0.2.0"
description = "FastAPI wrapper for Continuum SDK (demo + hosted MVP backend)"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "continuum-sdk>=0.1.1",
]

[project.optional-dependencies]
hosted = [
  "psycopg[binary]>=3.1",
  "bcrypt>=4.0",
  "PyJWT>=2.8",
]
